ironn/modules/dashboards/iCA/requirements.txt
dash==0.36.0
dash-html-components==0.13.5
dash-core-components==0.43.0
dash-renderer==0.17.0
numpy==1.16.3
pandas==0.24.2
plotly==2.7.0
flask==1.0.1
gunicorn==19.8.1
scipy==1.2.1
pillow==6.0.0
Flask-Caching==1.4.0
python-dotenv==0.8.2
matplotlib==3.1.0
PyYAML==5.1
